+++ blaze-core/src/main/scala/org/http4s/blazecore/websocket/Http4sWSStage.scala
@@ -39,7 +39,6 @@ import org.http4s.websocket.WebSocketFrame._
 
 import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success}
-import java.net.ProtocolException
 
 private[http4s] class Http4sWSStage[F[_]](
     ws: WebSocket[F],
