+++ client/jvm/src/main/scala/org/http4s/client/middleware/GZip.scala
