+++ client/jvm/src/test/scala/org/http4s/client/middleware/GZipSuite.scala
