+++ client/shared/src/main/scala/org/http4s/client/Connection.scala
