+++ client/shared/src/main/scala/org/http4s/client/DefaultClient.scala
