+++ client/shared/src/main/scala/org/http4s/client/Http4sClientDsl.scala
