+++ client/shared/src/main/scala/org/http4s/client/middleware/Retry.scala
