+++ client/shared/src/main/scala/org/http4s/client/middleware/package.scala
