+++ client/shared/src/main/scala/org/http4s/client/oauth1/Types.scala
