+++ client/shared/src/test/resources/testresource.txt
