+++ client/shared/src/test/scala/org/http4s/client/middleware/FollowRedirectSuite.scala
