+++ client/shared/src/test/scala/org/http4s/client/oauth1/SignatureAlgorithmSuite.scala
