--- core/jvm/src/main/scala/org/http4s/UrlEncoder.scala
@@ -1,26 +1,0 @@
-/*
- * Copyright 2013 http4s.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.http4s
-
-import java.net.URLEncoder
-import java.nio.charset
-
-private[http4s] object UrlEncoder {
-  private[this] val utf8 = charset.Charset.forName("UTF-8")
-
-  def encode(url: String): String = URLEncoder.encode(url, utf8)
-}
